refactor: Float로 받던 value를 Int로 변경

+++ app/src/main/java/com/sonchan/photoretouching/presentation/screen/RetouchingScreen.kt
@@ -125,7 +125,7 @@ fun RetouchingRoute(
         onToggleTheme = { themeViewModel.toggleTheme() },
         brightnessValue = brightnessValue,
         brightnessSliderState = brightnessSliderState,
-        onBrightnessChanged = { viewModel.updateRetouchingValue(RetouchingOption.BRIGHTNESS, it.toFloat()) },
+        onBrightnessChanged = { viewModel.updateRetouchingValue(RetouchingOption.BRIGHTNESS, it) },
         onBrightnessReset = { viewModel.resetRetouchingValue(RetouchingOption.BRIGHTNESS) }
     )
 }
